refactor: collapse feature api.ts modules into queries.ts
Fold each feature's per-endpoint api.ts request functions directly into
their React Query hooks and queryOptions/queryKey factories, removing the
separate @<feature>/api layer. Inline single-use requests into the hook's
queryFn/mutationFn; keep a module-private helper only where a request is
shared across hooks or branches (samples updateSample, uploads findFiles,
references member add/edit/remove).

Relocate types that lived in api.ts (sample update/rights types to
samples/types.ts; admin SettingsUpdate/UserUpdate, account AccountUpdate,
and wall LoginResult/ResetPasswordResult to their queries modules). Add
accountQueryOptions() and point the auth route loaders at it. Drop dead
exports (hmm/otus/references find/get, account get/settings helpers).
