+++ apps/web/src/hmm/queries.ts
@@ -1,3 +1,4 @@
+import { apiClient } from "@app/api";
 import {
 	keepPreviousData,
 	useMutation,
@@ -5,7 +6,6 @@ import {
 	useQueryClient,
 } from "@tanstack/react-query";
 import type { ErrorResponse } from "@/types/api";
-import { fetchHmm, installHmm, listHmms } from "./api";
 import type { HMMInstalled, Hmm, HmmSearchResults } from "./types";
 
 /**
@@ -31,7 +31,14 @@ export const hmmQueryKeys = {
 export function useListHmms(page: number, per_page: number, term?: string) {
 	return useQuery<HmmSearchResults>({
 		queryKey: hmmQueryKeys.list([page, per_page, term]),
-		queryFn: () => listHmms(page, per_page, term),
+		queryFn: () =>
+			apiClient
+				.get("/hmms")
+				.query({ page, per_page, find: term })
+				.then((res) => {
+					const { documents, ...rest } = res.body;
+					return { ...rest, items: documents };
+				}),
 		placeholderData: keepPreviousData,
 	});
 }
@@ -45,7 +52,7 @@ export function useListHmms(page: number, per_page: number, term?: string) {
 export function useFetchHmm(hmmId: string) {
 	return useQuery<Hmm, ErrorResponse>({
 		queryKey: hmmQueryKeys.detail(hmmId),
-		queryFn: () => fetchHmm(hmmId),
+		queryFn: () => apiClient.get(`/hmms/${hmmId}`).then((res) => res.body),
 		retry: (failureCount, error) => {
 			if (error.response?.status === 404) {
 				return false;
@@ -64,7 +71,8 @@ export function useInstallHmm() {
 	const queryClient = useQueryClient();
 
 	return useMutation<HMMInstalled, ErrorResponse>({
-		mutationFn: installHmm,
+		mutationFn: () =>
+			apiClient.post("/hmms/status/updates").then((res) => res.body),
 		onSuccess: () => {
 			queryClient.invalidateQueries({ queryKey: hmmQueryKeys.lists() });
 		},
