+++ apps/web/src/jobs/queries.ts
@@ -1,11 +1,12 @@
+import { apiClient } from "@app/api";
 import { keepPreviousData, useQuery } from "@tanstack/react-query";
-import { fetchJob, findJobs } from "./api";
 import {
 	JobSchema,
 	JobSearchResultSchema,
 	type JobState,
 	type ServerJob,
 	type ServerJobNested,
+	type ServerJobSearchResult,
 } from "./types";
 
 /**
@@ -35,7 +36,13 @@ export function useFindJobs(
 ) {
 	return useQuery({
 		queryKey: jobQueryKeys.list([page, per_page, ...states]),
-		queryFn: () => findJobs(page, per_page, states),
+		queryFn: async (): Promise<ServerJobSearchResult> => {
+			const response = await apiClient
+				.get("/jobs")
+				.query({ page, per_page, state: states });
+
+			return response.body;
+		},
 		placeholderData: keepPreviousData,
 		select: JobSearchResultSchema.parse,
 	});
@@ -74,7 +81,10 @@ function getJobSeed(job: ServerJobNested): ServerJob {
 export function useFetchJob(jobId: number, seed?: ServerJobNested) {
 	return useQuery({
 		queryKey: jobQueryKeys.detail(jobId),
-		queryFn: () => fetchJob(jobId),
+		queryFn: async (): Promise<ServerJob> => {
+			const response = await apiClient.get(`/jobs/${jobId}`);
+			return response.body;
+		},
 		select: JobSchema.parse,
 		enabled: Number.isInteger(jobId),
 		initialData: seed ? getJobSeed(seed) : undefined,
