+++ apps/web/src/routes/_authenticated/administration/index.tsx
@@ -1,5 +1,4 @@
-import { fetchAccount } from "@account/api";
-import { accountKeys } from "@account/queries";
+import { accountQueryOptions } from "@account/queries";
 import { hasSufficientAdminRole } from "@administration/utils";
 import { createFileRoute, redirect } from "@tanstack/react-router";
 
@@ -7,10 +6,7 @@ export const Route = createFileRoute("/_authenticated/administration/")({
 	beforeLoad: async ({ context }) => {
 		const { queryClient } = context;
 
-		const account = await queryClient.ensureQueryData({
-			queryKey: accountKeys.all(),
-			queryFn: fetchAccount,
-		});
+		const account = await queryClient.ensureQueryData(accountQueryOptions());
 
 		if (hasSufficientAdminRole("settings", account.administrator_role)) {
 			throw redirect({ to: "/administration/settings" });
