+++ apps/web/src/samples/types.ts
@@ -115,3 +115,32 @@ export type Sample = SampleMinimal & {
 export type SampleSearchResult = SearchResult & {
 	items: Array<SampleMinimal>;
 };
+
+/** Fields that can be changed when updating a sample */
+export type SampleUpdate = {
+	isolate?: string;
+	labels?: number[];
+	locale?: string;
+	name?: string;
+	notes?: string;
+	subtractions?: string[];
+};
+
+/** Data returned from API on sample rights update */
+export type SampleRightsUpdateReturn = {
+	all_read: boolean;
+	all_write: boolean;
+	group: number | string | null;
+	group_read: boolean;
+	group_write: boolean;
+	user: { [key: string]: string };
+};
+
+/** Fields that can be changed when updating a sample's rights */
+export type SampleRightsUpdate = {
+	group?: number | string | null;
+	group_read?: boolean;
+	group_write?: boolean;
+	all_read?: boolean;
+	all_write?: boolean;
+};
